refactor: bifurcate sim psu test to make for a better smoke test template
